Force alphanumeric sort on file list before import
os.walk returns files in an arbitary order, which can differ
between systems. This was causing the file load order to vary
on GH actions compared to my local Mac. As a result tests were
failing.

This change applies an alphanumeric sort to the files to ensure
they are loaded in a reproducible order. This does change the
previous arbitary behaviour that was being seen, and results in
older files appearing at the top of the combined file. This feels
less surprising, but we can do a sorted(..., reverse = True) if
we want the old order.