Remove FIELD_LAB_ID from required fields
The Mongo index will happily accept missing information for indexed fields.
So going to restore the removal of Lab ID as a required field. What this means:
 - Files containing Lab ID will be processed as normal, and lab ID will still be recorded in Mongo.
 - Files missing Lab ID will be processed. They will be recorded in the samples table with Lab ID clearly missing. (In compass it displays as ‘No Field’)
 - This meets the requirement to Parse with missing Lab ID and meets the general aim to be more permissive in what we accept, and to improve the visibility of missing information.
 - This avoids any risks of auto-populating the field based on centre, which could result in us recording incorrect information if one of the centres starts sending info from multiple labs.
It does mean that some reports using this information may not get immediate visibility on Lab ID, but they will have access to all the other information to begin chasing up any requirements.