Remove use of ABC as class to inherit
Without anything marked as an @abstractmethod this is useless and does nothing.  This is almost simple enough to be a mixin except that we use it as a return type in a number of places and is technically useful as a class in its own right.  Making it abstract would require us to have at least one method that we need sub-classes to override to be valid.